refactor: duplicated current verifier set query (#616)
